Refactor and adjust permission of user by role
