+++ src/discount/controllers/coupon.controller.ts
@@ -43,7 +43,7 @@ export class CouponController {
 
   @Post()
   @UseGuards(AuthGuard, RolesGuard)
-  @Roles(UserRole.ADMIN)
+  @Roles(UserRole.ADMIN, UserRole.BRANCH_ADMIN)
   @ApiOperation({ summary: 'Create a coupon' })
   @ApiResponse({
     status: HttpStatus.CREATED,
@@ -56,7 +56,7 @@ export class CouponController {
 
   @Get()
   @UseGuards(AuthGuard, RolesGuard)
-  @Roles(UserRole.ADMIN)
+  @Roles(UserRole.ADMIN, UserRole.BRANCH_ADMIN)
   @UseInterceptors(PaginationInterceptor)
   @ApiOperation({ summary: 'List all coupons' })
   @ApiQuery({
@@ -118,7 +118,10 @@ export class CouponController {
   }
 
   @HttpCode(HttpStatus.NO_CONTENT)
+  @UseGuards(AuthGuard, RolesGuard)
+  @Roles(UserRole.ADMIN, UserRole.BRANCH_ADMIN)
   @Delete('bulk')
+  @ApiBearerAuth()
   @ApiOperation({ summary: 'Bulk delete coupons' })
   @ApiResponse({
     description: 'Successful bulk deletion of coupons',
@@ -131,7 +134,10 @@ export class CouponController {
   }
 
   @HttpCode(HttpStatus.NO_CONTENT)
+  @UseGuards(AuthGuard, RolesGuard)
+  @Roles(UserRole.ADMIN, UserRole.BRANCH_ADMIN)
   @Patch('bulk')
+  @ApiBearerAuth()
   @ApiOperation({ summary: 'Bulk update coupons' })
   @ApiResponse({
     description: 'Successful bulk update of coupons',
@@ -160,7 +166,8 @@ export class CouponController {
 
   @Patch(':code')
   @UseGuards(AuthGuard, RolesGuard)
-  @Roles(UserRole.ADMIN)
+  @Roles(UserRole.ADMIN, UserRole.BRANCH_ADMIN)
+  @ApiBearerAuth()
   @ApiOperation({ summary: 'Update coupon by code' })
   @ApiResponse({
     status: HttpStatus.OK,
@@ -177,7 +184,8 @@ export class CouponController {
   @HttpCode(HttpStatus.NO_CONTENT)
   @Delete(':code')
   @UseGuards(AuthGuard, RolesGuard)
-  @Roles(UserRole.ADMIN)
+  @Roles(UserRole.ADMIN, UserRole.BRANCH_ADMIN)
+  @ApiBearerAuth()
   @ApiOperation({ summary: 'Delete coupon by code' })
   @ApiResponse({
     status: HttpStatus.NO_CONTENT,
