+++ src/order/services/order-delivery.controller.ts
@@ -1,0 +1,166 @@
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { User, UserRole } from 'src/user/entities/user.entity';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+import { OrderDelivery } from '../entities/order_delivery.entity';
+import { UpdateDeliveryDTO } from '../dto/order-delivery.dto';
+import { UserService } from 'src/user/user.service';
+
+@Injectable()
+export class OrderDeliveryService {
+  constructor(
+    @InjectRepository(OrderDelivery)
+    private orderDeliveryRepository: Repository<OrderDelivery>,
+    private userService: UserService,
+  ) {}
+
+  async findAll(
+    user: User,
+    page: number,
+    pageSize: number,
+    filters?: {
+      status?: string;
+      branchId?: string;
+      employeeId?: string;
+    },
+  ): Promise<OrderDelivery[]> {
+    const query = this.orderDeliveryRepository
+      .createQueryBuilder('delivery')
+      .leftJoinAndSelect('delivery.order', 'order')
+      .leftJoinAndSelect('order.user', 'orderUser')
+      .leftJoinAndSelect('orderUser.profile', 'profile')
+      .leftJoinAndSelect('delivery.address', 'address')
+      .leftJoinAndSelect('address.city', 'city')
+      .leftJoinAndSelect('city.state', 'state')
+      .leftJoinAndSelect('state.country', 'country')
+      .leftJoinAndSelect('delivery.branch', 'branch')
+      .leftJoinAndSelect('delivery.employee', 'employee')
+      .where('delivery.deletedAt IS NULL');
+
+    if (user.role !== UserRole.ADMIN) {
+      query.andWhere('employee.id = :userId', { userId: user.id });
+    } else if (filters?.employeeId) {
+      query.andWhere('employee.id = :employeeId', {
+        employeeId: filters.employeeId,
+      });
+    }
+
+    if (filters?.status) {
+      query.andWhere('delivery.deliveryStatus = :status', {
+        status: filters.status,
+      });
+    }
+
+    if (filters?.branchId) {
+      query.andWhere('branch.id = :branchId', { branchId: filters.branchId });
+    }
+
+    const delivery = query
+      .orderBy('delivery.createdAt', 'DESC')
+      .skip((page - 1) * pageSize)
+      .take(pageSize);
+
+    return await delivery.getMany();
+  }
+
+  async countDeliveries(
+    user: User,
+    filters?: {
+      status?: string;
+      branchId?: string;
+      employeeId?: string;
+    },
+  ): Promise<number> {
+    const query = this.orderDeliveryRepository
+      .createQueryBuilder('delivery')
+      .leftJoin('delivery.branch', 'branch')
+      .leftJoin('delivery.employee', 'employee')
+      .where('delivery.deletedAt IS NULL');
+
+    if (user.role !== UserRole.ADMIN) {
+      query.andWhere('employee.id = :userId', { userId: user.id });
+    } else if (filters?.employeeId) {
+      query.andWhere('employee.id = :employeeId', {
+        employeeId: filters.employeeId,
+      });
+    }
+
+    if (filters?.status) {
+      query.andWhere('delivery.deliveryStatus = :status', {
+        status: filters.status,
+      });
+    }
+
+    if (filters?.branchId) {
+      query.andWhere('branch.id = :branchId', { branchId: filters.branchId });
+    }
+
+    return await query.getCount();
+  }
+
+  async findOne(deliveryId: string): Promise<OrderDelivery> {
+    const delivery = await this.orderDeliveryRepository.findOne({
+      where: { id: deliveryId },
+      relations: [
+        'order',
+        'order.user',
+        'order.user.profile',
+        'address',
+        'address.city',
+        'address.city.state',
+        'address.city.state.country',
+        'employee',
+        'branch',
+      ],
+    });
+    if (!delivery) {
+      throw new NotFoundException('Delivery not found.');
+    }
+    return delivery;
+  }
+
+  async update(
+    user: User,
+    deliveryId: string,
+    updateData: UpdateDeliveryDTO,
+  ): Promise<OrderDelivery> {
+    const delivery = await this.orderDeliveryRepository.findOne({
+      where: { id: deliveryId },
+      relations: [
+        'order',
+        'order.user',
+        'address',
+        'address.city',
+        'address.city.state',
+        'address.city.state.country',
+        'employee',
+        'branch',
+      ],
+    });
+    if (!delivery) {
+      throw new NotFoundException('Delivery not found.');
+    }
+
+    const updateDelivery = this.orderDeliveryRepository.merge(
+      delivery,
+      updateData,
+    );
+    if (updateData.employeeId) {
+      const employee = await this.userService.findUserById(
+        updateData.employeeId,
+      );
+      if (!employee) {
+        throw new NotFoundException('Employee not found.');
+      }
+      if (employee.role !== UserRole.DELIVERY) {
+        throw new BadRequestException('User is not an employee.');
+      }
+      updateDelivery.employee = employee;
+    }
+    return await this.orderDeliveryRepository.save(updateDelivery);
+  }
+}
